// src/types/game.ts
export interface FirePuddle {
  id: string;
  x: number;
  y: number;
  radius: number;
  duration: number;
  createdTime: number;
}